requirements.txt
pdfplumber
openai
pinecone-client
python-dotenv
spacy
pytesseract
pillow
pymupdf
